refacto(services): separate client & server services
