+++ app/components/client/tasks/task.component.tsx
@@ -1,7 +1,7 @@
 import { Task } from "@/app/types/task.type";
 import TaskCheckbox from "../checkbox.component";
 import DeleteIcon from '@mui/icons-material/Delete';
-import { TasksDataClientService } from "@/app/services/tasks-data-client.service";
+import { TasksDataClientService } from "@/app/services/client/tasks-data-client.service";
 
 export default function TaskItem ({task, setTasks}: {task: Task, setTasks: (tasks: Task[]) => unknown}) {
